% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_alg.R
\name{online_alg}
\alias{online_alg}
\title{Online Algorithm for Renewable Expectile Regression}
\usage{
online_alg(X, Y, Xc, Yc, K, tau, regFormula, dataName)
}
\arguments{
\item{X}{A list of covariate matrices from the streaming data, where each element corresponds to the covariates of a single data batch.}

\item{Y}{A list of response vectors from the streaming data, where each element corresponds to the response in a single data batch.}

\item{Xc}{Centralized version of \code{X}.}

\item{Yc}{Centralized version of \code{Y}.}

\item{K}{Integer. Number of data batches.}

\item{tau}{Numeric. Expectile level (between 0 and 1).}

\item{regFormula}{A formula specifying the regression model, e.g., \code{Y ~ x1 + x2}.}

\item{dataName}{Character vector. Variable names in the data, e.g., \code{c("Y", "x1", "x2")}.}
}
\value{
A list with three components:
\describe{
  \item{DCER}{Distributed estimation via precision matrix averaging.}
  \item{PAER}{Parameter aggregation using weighted local estimators.}
  \item{ReER}{Renewable estimation using Taylor approximation.}
}
}
\description{
Implements an online (streaming) algorithm for expectile regression using different renewable updating strategies.
}
\details{
The function iteratively updates regression parameters for streaming data using different online algorithms. Expectile regression is estimated via the \code{expectreg.ls} function in each batch.
}
\examples{
library(expectreg)
library(MASS)
library(dplyr)
library(ReER)

# generate data
N <- 1000; k <- 10
x1 <- runif(N); x2 <- runif(N)
Xmat <- cbind(x1, x2)
X <- cbind(1, Xmat); Xc <- cbind(1, scale(Xmat, scale = FALSE))
u <- rnorm(N)
beta <- c(2,1,2); gamma <- c(1, 0, 0)
Y <- X \%*\% beta + X \%*\% gamma * u
Yc <- Xc \%*\% beta + Xc \%*\% gamma * u

# split data
indices <- split(1:N, cut(1:N, breaks = k, labels = FALSE))
X_split <- lapply(indices, function(idx) X[idx, , drop = FALSE])
Y_split <- lapply(indices, function(idx) Y[idx])
Xc_split <- lapply(indices, function(idx) Xc[idx, , drop = FALSE])
Yc_split <- lapply(indices, function(idx) Yc[idx])

result <- online_alg(X = X_split, Y = Y_split,
                     Xc = Xc_split, Yc = Yc_split,
                     K = k, tau = 0.25,
                     regFormula = Y ~ x1 + x2,
                     dataName = c("Y", "x0", "x1", "x2"))
}
