% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_alg.R
\name{qMatrix}
\alias{qMatrix}
\title{Calculate the asymptotic covariance matrix of expectile regression}
\usage{
qMatrix(X, Y, beta, tau)
}
\arguments{
\item{X}{Covariates matrices.}

\item{Y}{Independent vector.}

\item{beta}{Parameters.}

\item{tau}{Expectile level.}
}
\value{
The asymptotic covariance matrix of expectile regression.
}
\description{
Calculate the asymptotic covariance matrix of expectile regression
}
