% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_alg.R
\name{ReER}
\alias{ReER}
\title{Renewable estimation for expectile regression}
\usage{
ReER(X, Y, beta, Z_cum, tau)
}
\arguments{
\item{X}{A covaraites matrix.}

\item{Y}{A vector object.}

\item{beta}{Initial estimation of parameters.}

\item{Z_cum}{The cumulative Hessian matrix for history data.}

\item{tau}{Expectile level.}
}
\value{
A list object, which is basically a list consisting of:
\item{beta_new}{Renewable estimator}
\item{Z_cum}{Updated cumulative Hessian matrix}
\item{intbeta}{The update imputation parameters}
}
\description{
Renewable estimation derived with Taylor approximation.
}
