% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_alg.R
\name{weightMatrix}
\alias{weightMatrix}
\title{Calculate the diagnol element of expectile weight matrix}
\usage{
weightMatrix(tau, X, Y, beta)
}
\arguments{
\item{tau}{Expectile level.}

\item{X}{Covariates matrices.}

\item{Y}{Independent vector.}

\item{beta}{Parameters.}
}
\value{
Expectile weight vector.
}
\description{
Calculate the diagnol element of expectile weight matrix
}
